pyproject.toml
[project]
name = "voice-portfolio"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "httpx>=0.27.0",
    "pydantic-settings>=2.12.0",
]
